Got rid of shorts

+++ src/gui/Constants.java
@@ -19,10 +19,10 @@ public interface Constants {
 	public static final int Error_CRC = 0x22;
 	public static final int Error_Read = 0x23;
 	public static final String MD_MANAGER_CONFIG = System.getProperty("user.home") + System.getProperty("file.separator") + "megadrummanager.cfg";
-	public static final short PADS_COUNT = 55;
-	public static final short CURVES_COUNT = 4;
-	public static final short CUSTOM_NAMES_MAX = 32;
-	public static final short SYSEX_DELAY = 50;
+	public static final int PADS_COUNT = 55;
+	public static final int CURVES_COUNT = 4;
+	public static final int CUSTOM_NAMES_MAX = 32;
+	public static final int SYSEX_DELAY = 50;
 	public static final byte SYSEX_START = (byte)0xf0;
 	public static final byte SYSEX_END = (byte)0xf7;
 	public static final byte MD_SYSEX = (byte)0x70;
