+++ src/gui/Spin127Control.java
@@ -51,11 +51,11 @@ public class Spin127Control extends JPanel {
 		spinner.setModel(model);
 	}
 	
-	public short getValue() {
-		return ((Short)spinner.getValue()).shortValue();
+	public int getValue() {
+		return (Integer)spinner.getValue();
 	}
 	
-	public void setValue(short value) {
+	public void setValue(int value) {
 		if ((value >= 0 ) && (value < 128)) {
 			spinner.setValue(value);
 		}
